requirements.txt
customtkinter==5.2.1
pynput==1.7.6
Pillow==10.2.0
requests==2.31.0